+++ app/Http/Controllers/UserChangePassController.php
@@ -64,9 +64,8 @@ class UserChangePassController extends Controller
         $last_img = 'profile-photos/'.$name_gen;
         Image::make($user_image)->fit(400,400)->save('storage/'.$last_img);
 
-        if($old_image) {
-            unlink('storage/'.$old_image);
-        }
+        if(file_exists($old_image)) unlink('storage/'.$old_image);
+        
         $user->profile_photo_path = $last_img; 
         }
         $user->name = $request->name;
